Reorganize tests and break them down into separate asmdef files
