+++ unity/PitayaExample/Assets/Tests/Integration.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 762be7ca838443bcb61cf212aeef19e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
