+++ unity/PitayaExample/Assets/Tests/Integration/JsonSSLRequestTest.cs
@@ -7,25 +7,35 @@ using Pitaya.SimpleJson;
 
 namespace Pitaya.Tests
 {
-    public class JsonRequestTest
+    [TestFixture, Category("integration")]
+    public class JsonSSLRequestTest
     {
+
         private PitayaClient _client;
 
         private const string ServerHost = "libpitaya-tests.tfgco.com";
-        private const int ServerPort = 3251;
+
+        private const int ServerPort = 3252;
         private JsonObject _jsonStub;
         private JsonObject _emptyJsonStub;
         private bool _isFinished;
         private string _data;
         private string _emptyData;
+        private Thread _mainThread;
 
         [SetUp]
         public void Init()
         {
-            _client = new PitayaClient();
+            // PitayaBinding.AddPinnedPublicKeyFromCaFile("client-ssl.localhost.crt");
+            _mainThread = Thread.CurrentThread;
+            PitayaBinding.SkipKeyPinCheck(true);
+
+            _client = new PitayaClient("myCA.pem");
             PitayaClient.SetLogLevel(PitayaLogLevel.Debug);
 
             _client.Connect(ServerHost, ServerPort);
+
+            Thread.Sleep(200);
 
             _data = "{\"Data\":{\"name\":\"test25\"}}";
             _jsonStub = (JsonObject)SimpleJson.SimpleJson.DeserializeObject(_data);
@@ -34,18 +44,12 @@ namespace Pitaya.Tests
 
             _isFinished = false;
 
-            Thread.Sleep(300);
-
             // clearing sessiondata
             _client.Request(
                 "connector.setsessiondata",
                 _emptyData,
-                res => {
-                    _isFinished = true;
-                },
-                error => {
-                    _isFinished = true;
-                }
+                res => { _isFinished = true; },
+                error => { _isFinished = true; }
             );
         }
 
@@ -61,9 +65,11 @@ namespace Pitaya.Tests
         [UnityTest]
         public IEnumerator ShouldReturnJsonObjectWhenNoSubscriberIsAdded()
         {
-            while(!_isFinished) {
+            while (!_isFinished)
+            {
                 yield return new WaitForSeconds(0.3f);
             }
+
             _isFinished = false;
 
             object response = null;
@@ -74,18 +80,20 @@ namespace Pitaya.Tests
                 res => {
                     response = res;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 },
                 error => {
                     response = error;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 }
             );
 
             while(!_isFinished) {
-                yield return new WaitForSeconds(0.3f);
+                yield return new WaitForSeconds(1);
             }
 
-            var resp = (JsonObject)SimpleJson.SimpleJson.DeserializeObject((string)response);
+            var resp = (JsonObject)SimpleJson.SimpleJson.DeserializeObject((string) response);
 
             Assert.True(_isFinished);
             Assert.NotNull(response);
@@ -96,9 +104,11 @@ namespace Pitaya.Tests
         [UnityTest]
         public IEnumerator ShouldGetAndSetJsonObject()
         {
-            while(!_isFinished) {
+            while (!_isFinished)
+            {
                 yield return new WaitForSeconds(0.3f);
             }
+
             _isFinished = false;
 
             object response = null;
@@ -109,15 +119,17 @@ namespace Pitaya.Tests
                 res => {
                     response = res;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 },
                 error => {
                     response = error;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 }
             );
 
             while(!_isFinished) {
-                yield return new WaitForSeconds(0.3f);
+                yield return new WaitForSeconds(1);
             }
 
             _isFinished = false;
@@ -128,29 +140,36 @@ namespace Pitaya.Tests
                 res => {
                     response = res;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 },
                 error => {
                     response = error;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 }
             );
 
             while(!_isFinished) {
-                yield return new WaitForSeconds(0.3f);
+                yield return new WaitForSeconds(1);
             }
 
             Assert.True(_isFinished);
             Assert.NotNull(response);
-            Assert.AreEqual(_data, response);
-            Assert.AreEqual(_jsonStub, (JsonObject)SimpleJson.SimpleJson.DeserializeObject((string)response));
+            Assert.AreEqual(response, _data);
+            Assert.AreEqual(SimpleJson.SimpleJson.DeserializeObject((string)response), _jsonStub);
+
+            _isFinished = false;
+
         }
 
         [UnityTest]
         public IEnumerator ShouldReceiveEmptyDataWhenSettingNullRequest()
         {
-            while(!_isFinished) {
+            while (!_isFinished)
+            {
                 yield return new WaitForSeconds(0.3f);
             }
+
             _isFinished = false;
 
             object response = null;
@@ -160,15 +179,17 @@ namespace Pitaya.Tests
                 res => {
                     response = res;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 },
                 error => {
                     response = error;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 }
             );
 
             while(!_isFinished) {
-                yield return new WaitForSeconds(0.3f);
+                yield return new WaitForSeconds(0.5f);
             }
 
             _isFinished = false;
@@ -179,29 +200,33 @@ namespace Pitaya.Tests
                 res => {
                     response = res;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 },
                 error => {
                     response = error;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 }
             );
 
             while(!_isFinished) {
-                yield return new WaitForSeconds(0.3f);
+                yield return new WaitForSeconds(0.5f);
             }
 
             Assert.True(_isFinished);
             Assert.NotNull(response);
-            var responseJson = (JsonObject) SimpleJson.SimpleJson.DeserializeObject((string) response);
-            Assert.True(responseJson.ContainsKey("Data"));
+            Assert.AreEqual(response, _emptyData);
+            Assert.AreEqual(SimpleJson.SimpleJson.DeserializeObject((string)response), _emptyJsonStub);
         }
 
         [UnityTest]
         public IEnumerator ShouldReturnBadRequestWhenRequestingToInvalidRoute()
         {
-            while(!_isFinished) {
+            while (!_isFinished)
+            {
                 yield return new WaitForSeconds(0.3f);
             }
+
             _isFinished = false;
 
             object response = null;
@@ -211,15 +236,17 @@ namespace Pitaya.Tests
                 res => {
                     response = res;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 },
                 error => {
                     response = error;
                     _isFinished = true;
+                    Assert.AreEqual(_mainThread, Thread.CurrentThread);
                 }
             );
 
             while(!_isFinished) {
-                yield return new WaitForSeconds(0.3f);
+                yield return new WaitForSeconds(0.5f);
             }
 
             Assert.True(_isFinished);
