+++ unity/PitayaExample/Assets/Tests/Integration/Tests.Integration.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "Tests",
+    "name": "Tests.Integration",
     "references": [
         "Pitaya",
         "UnityEngine.TestRunner",
